+++ cmd/courier/main.go
@@ -25,6 +25,7 @@ import (
 	_ "github.com/nyaruka/courier/handlers/globe"
 	_ "github.com/nyaruka/courier/handlers/highconnection"
 	_ "github.com/nyaruka/courier/handlers/hub9"
+	_ "github.com/nyaruka/courier/handlers/i2sms"
 	_ "github.com/nyaruka/courier/handlers/infobip"
 	_ "github.com/nyaruka/courier/handlers/jasmin"
 	_ "github.com/nyaruka/courier/handlers/jiochat"