+++ handlers/whatsapp/whatsapp.go
@@ -334,17 +334,19 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 	mediaURL := url.ResolveReference(mediaPath).String()
 
 	status := h.Backend().NewMsgStatusForID(msg.Channel(), msg.ID(), courier.MsgErrored)
+	var log *courier.ChannelLog
 
 	if len(msg.Attachments()) > 0 {
 		for attachmentCount, attachment := range msg.Attachments() {
 
 			mimeType, s3url := handlers.SplitAttachment(attachment)
-			mediaID, err := uploadMediaToWhatsApp(mediaURL, token, mimeType, s3url)
+			mediaID := ""
+			mediaID, log, err = uploadMediaToWhatsApp(msg, mediaURL, token, mimeType, s3url)
+			status.AddLog(log)
+
 			if err != nil {
-				duration := time.Now().Sub(start)
-				log := courier.NewChannelLogFromError("Unable to upload media to WhatsApp server", msg.Channel(), msg.ID(), duration, err)
-				status.AddLog(log)
-				return status, err
+				log.WithError("Unable to upload media to WhatsApp server", err)
+				break
 			}
 
 			externalID := ""
@@ -354,7 +356,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 					Type: "audio",
 				}
 				payload.Audio = &mediaObject{ID: mediaID}
-				externalID, err = sendWhatsAppMsg(sendURL, token, payload)
+				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
 
 			} else if strings.HasPrefix(mimeType, "application") {
 				payload := mtDocumentPayload{
@@ -367,7 +369,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				} else {
 					payload.Document = &captionedMediaObject{ID: mediaID}
 				}
-				externalID, err = sendWhatsAppMsg(sendURL, token, payload)
+				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
 
 			} else if strings.HasPrefix(mimeType, "image") {
 				payload := mtImagePayload{
@@ -379,26 +381,33 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				} else {
 					payload.Image = &captionedMediaObject{ID: mediaID}
 				}
-				externalID, err = sendWhatsAppMsg(sendURL, token, payload)
+				externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
 
 			} else {
+				duration := time.Since(start)
 				err = fmt.Errorf("unknown attachment mime type: %s", mimeType)
+				log = courier.NewChannelLogFromError("Error sending message", msg.Channel(), msg.ID(), duration, err)
 			}
 
-			if err != nil {
-				// record our status and log
-				duration := time.Now().Sub(start)
-				log := courier.NewChannelLogFromError("Error sending message", msg.Channel(), msg.ID(), duration, err)
+			// if we have a log, add it to our status
+			if log != nil {
 				status.AddLog(log)
-				return status, err
 			}
 
-			status.SetExternalID(externalID)
+			// break out on errors
+			if err != nil {
+				break
+			}
 
+			// set our external id if we have one
+			if attachmentCount == 0 {
+				status.SetExternalID(externalID)
+			}
 		}
 
 	} else {
 		parts := handlers.SplitMsg(msg.Text(), maxMsgLength)
+		externalID := ""
 		for i, part := range parts {
 			payload := mtTextPayload{
 				To:   msg.URN().Path(),
@@ -406,13 +415,12 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 			}
 			payload.Text.Body = part
 
-			externalID, err := sendWhatsAppMsg(sendURL, token, payload)
+			externalID, log, err = sendWhatsAppMsg(msg, sendURL, token, payload)
+			status.AddLog(log)
+
 			if err != nil {
-				// record our status and log
-				duration := time.Now().Sub(start)
-				log := courier.NewChannelLogFromError("Error sending message", msg.Channel(), msg.ID(), duration, err)
-				status.AddLog(log)
-				return status, err
+				log.WithError("Error sending message", err)
+				break
 			}
 
 			// if this is our first message, record the external id
@@ -423,17 +431,20 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 
 	}
 
-	status.SetStatus(courier.MsgWired)
+	// we are wired it there were no errors
+	if err == nil {
+		status.SetStatus(courier.MsgWired)
+	}
+
 	return status, nil
 }
 
-func uploadMediaToWhatsApp(url string, token string, attachmentMimeType string, attachmentURL string) (string, error) {
-
+func uploadMediaToWhatsApp(msg courier.Msg, url string, token string, attachmentMimeType string, attachmentURL string) (string, *courier.ChannelLog, error) {
 	// retrieve the media to be sent from S3
 	req, _ := http.NewRequest(http.MethodGet, attachmentURL, nil)
 	s3rr, err := utils.MakeHTTPRequest(req)
 	if err != nil {
-		return "", err
+		return "", courier.NewChannelLogFromRR("Media Fetch", msg.Channel(), msg.ID(), s3rr), err
 	}
 
 	// upload it to WhatsApp in exchange for a media id
@@ -442,23 +453,26 @@ func uploadMediaToWhatsApp(url string, token string, attachmentMimeType string, 
 	waReq.Header.Set("Content-Type", attachmentMimeType)
 	waReq.Header.Set("User-Agent", utils.HTTPUserAgent)
 	wArr, err := utils.MakeHTTPRequest(waReq)
+
+	log := courier.NewChannelLogFromRR("Media Upload success", msg.Channel(), msg.ID(), wArr)
+
 	if err != nil {
-		return "", err
+		return "", log, err
 	}
 
 	mediaID, err := jsonparser.GetString(wArr.Body, "media", "[0]", "id")
 	if err != nil {
-		return "", err
+		return "", log, err
 	}
 
-	return mediaID, nil
+	return mediaID, log, nil
 }
 
-func sendWhatsAppMsg(url string, token string, payload interface{}) (string, error) {
-
+func sendWhatsAppMsg(msg courier.Msg, url string, token string, payload interface{}) (string, *courier.ChannelLog, error) {
 	jsonBody, err := json.Marshal(payload)
 	if err != nil {
-		return "", err
+		log := courier.NewChannelLog("unable to build JSON body", msg.Channel(), msg.ID(), "", "", courier.NilStatusCode, "", "", time.Duration(0), err)
+		return "", log, err
 	}
 
 	req, _ := http.NewRequest(http.MethodPost, url, bytes.NewReader(jsonBody))
@@ -468,18 +482,20 @@ func sendWhatsAppMsg(url string, token string, payload interface{}) (string, err
 	req.Header.Set("User-Agent", utils.HTTPUserAgent)
 	rr, err := utils.MakeHTTPRequest(req)
 
+	log := courier.NewChannelLogFromRR("Message Sent", msg.Channel(), msg.ID(), rr).WithError("Message Send Error", err)
+
 	errorTitle, err := jsonparser.GetString(rr.Body, "errors", "[0]", "title")
 	if errorTitle != "" {
 		err = errors.Errorf("received error from send endpoint: %s", errorTitle)
-		return "", err
+		return "", log, err
 	}
 
 	// grab the id
 	externalID, err := jsonparser.GetString(rr.Body, "messages", "[0]", "id")
 	if err != nil {
 		err := errors.Errorf("unable to get message id from response body")
-		return "", err
+		return "", log, err
 	}
 
-	return externalID, err
+	return externalID, log, err
 }